+++ feature/work/data/src/main/java/com/andef/mycarandef/work/data/dbo/WorkDbo.kt
@@ -27,5 +27,5 @@ data class WorkDbo(
     val mileage: Int,
     val date: Int,
     @ColumnInfo(name = "car_id")
-    val carId: Int
+    val carId: Long
 )
